Changed the way we draw to the screen
Instead of using the spriteBatch object, I switched to using the Stage
object. The stage object is built into the API and it already uses a
camera and spriteBatch in its implementation, AND it allows us to add
buttons and pictures pretty easily. From now on we should pass the stage
object into the constructors of new classes instead of a spriteBatch.